Account for "holes" in the existing migrations in PlanMigration
If there are migrations missing before the last run migration, for
example because of a merge, migrate up to the latest run migration
first before applying the user-requested migrations.

Example: if the existing migrations in the database are:

- 1
- 3

and the existing migrations on disk are:

- 1
- 2
- 3
- 4
- 5

For "up, n=0" (3=>5) one would expect these migrations to run:
- up 2
- up 4
- up 5

For "up, n=1" (3=>4) one would expect:
- up 2
- up 4

For "down, n=1" (3=>2) one would expect:
- up 2
- down 3
- down 2

etc.