Return error from PlanMigration when unknown migration in the database
As discussed in GitHub issue #43 the PlanMigration function would not
allow running migrations when the there are already applied migrations
in the database which are not part of the migration source. In such
cases the user have to decide what to do.

This also fixes a bug where running the `up` migration would result in
success exit status while nothing has been done.