Fallback to reading Environment from shell Env
We want to build the ability for this command to accept the Environment
through the shell env. If the file is not present, we optimistically
look for the Environment in said shell env. There are existing
zero-value checks to make sure that a user doesn't accidentally end up
with no values for anything if their file was empty. We do a separate
stat to check for the file to keep the complexity of the fallback
simpler.

The envconfig package can be researched more here:

https://github.com/kelseyhightower/envconfig